requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]
inngest

# LangChain and LLM (required by llm_integration module)
langchain-core
langchain-mistralai
langchain-redis

# Other project dependencies
python-dotenv
requests
streamlit

# Database and Data Storage
pydantic
pinecone
pymongo
motor

ragas==0.1.21
chromadb
sqlalchemy
psycopg2-binary
redis

# Evaluation and Monitoring
ragas==0.1.21
clearml

# Data Science and ML
numpy
torch
torchvision
